fix header pos error
