+++ paddle/phi/kernels/cpu/affine_grid_grad_kernel.cc
@@ -13,9 +13,11 @@
 // limitations under the License.
 
 #include "paddle/phi/kernels/affine_grid_grad_kernel.h"
+
 #include "paddle/phi/backends/cpu/cpu_context.h"
 #include "paddle/phi/common/int_array.h"
 #include "paddle/phi/core/kernel_registry.h"
+#include "paddle/phi/kernels/funcs/affine_grid_utils.h"
 
 namespace phi {
