+++ paddle/phi/kernels/cpu/fill_diagonal_grad_kernel.cc
@@ -15,6 +15,7 @@
 #include "paddle/phi/kernels/fill_diagonal_grad_kernel.h"
 
 #include "paddle/phi/core/kernel_registry.h"
+#include "paddle/phi/kernels/funcs/common_shape.h"
 
 namespace phi {
