+++ paddle/phi/kernels/cpu/fill_diagonal_kernel.cc
@@ -16,6 +16,7 @@
 
 #include "paddle/phi/core/kernel_registry.h"
 #include "paddle/phi/core/tensor_utils.h"
+#include "paddle/phi/kernels/funcs/common_shape.h"
 
 namespace phi {
