+++ paddle/phi/kernels/fill_diagonal_grad_kernel.h
@@ -16,8 +16,6 @@
 
 #include "paddle/phi/core/dense_tensor.h"
 
-#include "paddle/phi/kernels/impl/fill_diagonal_kernel_impl.h"
-
 namespace phi {
 
 template <typename T, typename Context>
