+++ paddle/phi/kernels/fill_diagonal_kernel.h
@@ -15,9 +15,6 @@
 #pragma once
 
 #include "paddle/phi/core/dense_tensor.h"
-
-#include "paddle/phi/kernels/impl/fill_diagonal_kernel_impl.h"
-
 namespace phi {
 
 template <typename T, typename Context>
