+++ paddle/phi/kernels/gpu/fill_diagonal_grad_kernel.cu
@@ -18,6 +18,7 @@
 #include <vector>
 
 #include "paddle/phi/core/kernel_registry.h"
+#include "paddle/phi/kernels/funcs/common_shape.h"
 
 namespace phi {
